- extract styles and mock data from restaurant-list component to separate files
