+++ src/features/restaurants/components/restaurant-list/mock.data.js
@@ -1,0 +1,26 @@
+export const dummyRestaurantsData = [
+  {
+    restaurant_name: "Là Appetit",
+    restaurant_icon:
+      "https://www.deputy.com/uploads/2018/10/The-Most-Popular-Menu-Items-That-You-should-Consider-Adding-to-Your-Restaurant_Content-image1-min-1024x569.png",
+    restaurant_photos: [
+      "https://www.deputy.com/uploads/2018/10/The-Most-Popular-Menu-Items-That-You-should-Consider-Adding-to-Your-Restaurant_Content-image1-min-1024x569.png",
+    ],
+    restaurant_address: "165 Polk Street, San Francisco",
+    isOpen: true,
+    restaurant_rating: 5,
+    isTemporarilyClosed: false,
+  },
+  {
+    restaurant_name: "Zuni Café",
+    restaurant_icon:
+      "https://images.pexels.com/photos/1640772/pexels-photo-1640772.jpeg?cs=srgb&dl=pexels-ella-olsson-1640772.jpg&fm=jpg",
+    restaurant_photos: [
+      "https://images.pexels.com/photos/1640772/pexels-photo-1640772.jpeg?cs=srgb&dl=pexels-ella-olsson-1640772.jpg&fm=jpg",
+    ],
+    restaurant_address: "1658 Markte Street, San Francisco",
+    isOpen: false,
+    restaurant_rating: 3,
+    isTemporarilyClosed: true,
+  },
+];
