+++ src/features/restaurants/components/restaurant-list/restaurant-list.styles.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const RestaurantListContainer = styled.View`
+  display: flex;
+  flex: 1;
+`;
+
+export const RestaurantDataList = styled.View`
+  display: flex;
+  flex: 1;
+  padding: ${(props) => props.theme.space[2]};
+`;
